requirements.txt
# Core dependencies for the gemini CLI hack project
pypdf>=3.0.0
Pillow>=9.0.0
requests>=2.28.0
# Note: requests needed for OpenAI DALL-E API calls for image generation
